src/main/java/recap/v2/dp/LongestCommonSubsequence.java
/*
 * Created by AdelaSarca
 * Date: 2/27/2020
 */
package recap.v2.dp;

import java.util.LinkedList;

public class LongestCommonSubsequence {

    public String common(String s1, String s2) {
        int n = s1.length();
        int m = s2.length();

        int[][] dp = new int[n + 1][m + 1];
        for (int i = 1; i <= n; i++) {
            for (int j = 1; j <= m; j++) {
                if (s1.charAt(i - 1) == s2.charAt(j - 1)) {
                    dp[i][j] = 1 + dp[i - 1][j - 1];
                } else {
                    dp[i][j] = Integer.max(dp[i - 1][j], dp[i][j - 1]);
                }
            }
        }

        StringBuilder solution = new StringBuilder();
        int i = n, j = m;
        while(i > 0 && j > 0) {
            if (s1.charAt(i - 1) == s2.charAt(j - 1)) {
                solution.insert(0, s1.charAt(i - 1));
                i--;
                j--;
            } else if (dp[i][j - 1] > dp[i - 1][j]) {
                j--;
            } else {
                i--;
            }
        }
        return solution.toString();
    }
}
